0_Garbage/requirements.txt
Flask==2.3.2
Pillow==9.5.0
torch==2.4.1
torchvision==0.19.1  # Update this line to use the latest available version
transformers==4.30.0